requirements.txt
adaptivecards==0.4.1
aiohttp==3.8.6
azure_storage==0.37.0
botbuilder_core==4.16.2
botbuilder_schema==4.16.2
openai==0.28.0
python-dotenv==1.0.1